Correct player leave stat recalc
Currently, when OnPlayerLeaveAll is called, the player is still counted in GetPlayersCountExceptGMs so switching to a regular decrements, as it is called once per player leaving the map, allows for it to properly recalculate the stats of monsters as players leave the map. This commit also changes the instance of GetPlayersCountExceptGMs in OnPlayerEnterAll to a simple increment as it is called once per player entering the map as well.