Level scaling is broken because the OnCreateMap is called after OnBeforeCreatureSelectLevel
This lead to lfg level information not being set when the map is created, and the level scaling logic
is not able to find the lfg level information. This commit fixes that by moving the logic from
OnCreateMap to a new function called GetMapInfo used in all the places where we need to get the
autobalance map info. The first time the map is requested, it will be initialized and the data will be set.